test(build): split the sync test module at the fixture seam
The test module crossed 500 lines and became the one taste violation that
raised the count ratchet to 576 against a baseline of 575. Measured after
merging origin/main, per ci-scripts MUST 14: the branch alone reported 573,
the merged tree 576, and the merge is what makes the two comparable.

The seam is the fixture. Everything that drives the temporary tree stays in
test_sync_slim_agents.py. The delimiter parsing, which is pure functions, and
the shipped-constant checks, which must not see the fixture's monkeypatched
constants, move to test_sync_slim_agents_contract.py. That follows the
existing test_generate_hooks_*.py convention of one concern per module rather
than a suppression comment.

The new module loads the script under its own sys.modules name, so neither
module can displace the other's instance and share a patched constant.

Ratchet back to 575, equal to baseline. 48 cases still pass, and inverting
the sync direction still turns test_shipped_source_is_the_canonical_claude_tree
red from its new home.

Refs #5282
Refs #5526
